[ADD] Positions of IObject are now floats

+++ include/Object.hpp
@@ -35,22 +35,22 @@ namespace Arcade
         unsigned char m_symbol;
         Arcade::Color m_color;
         // For Both;
-        std::pair<unsigned int, unsigned int> m_pos;
+        std::pair<float, float> m_pos;
     public:
-        Tile(std::string path_to_bmp, unsigned char symbol, Arcade::Color color = WHITE, unsigned int x = 0, unsigned int y = 0)
+        Tile(std::string path_to_bmp, unsigned char symbol, Arcade::Color color = WHITE, float x = 0, float y = 0)
                 : m_path(std::move(path_to_bmp)), m_rotation(0), m_symbol(symbol), m_color(color), m_pos(std::make_pair(x, y)) {};
         ~Tile() override = default;
         // GETTERS
         [[nodiscard]] std::string getPath() const {return m_path;};
         [[nodiscard]] unsigned char getSymbol() const {return m_symbol;};
         [[nodiscard]] Arcade::Color getColor() const {return m_color;};
-        [[nodiscard]] std::pair<unsigned int, unsigned int> getPosition() const {return m_pos;};
+        [[nodiscard]] std::pair<float, float> getPosition() const {return m_pos;};
         [[nodiscard]] unsigned int getRotation() const {return m_rotation;};
         // SETTERS
         void setPath(std::string path_to_bmp) {m_path = std::move(path_to_bmp);};
         void setSymbol(unsigned char symbol) {m_symbol = symbol;};
         void setColor(Arcade::Color color) {m_color = color;};
-        void setPosition(unsigned int x, unsigned int y) {m_pos.first = x; m_pos.second = y;};
+        void setPosition(float x, float y) {m_pos.first = x; m_pos.second = y;};
         void setRotation(unsigned int angle) {m_rotation = angle;};
     };
 
@@ -67,11 +67,11 @@ namespace Arcade
         {m_tiles.push_back(main_tile);};
         ~DynamicTile() override = default;
         // GETTERS
-        [[nodiscard]] std::pair<unsigned int, unsigned int> getPosition() const {return m_tiles[m_iterator].getPosition();};
+        [[nodiscard]] std::pair<float, float> getPosition() const {return m_tiles[m_iterator].getPosition();};
         [[nodiscard]] unsigned int getRotation() const {return m_tiles[m_iterator].getRotation();};
         Arcade::Tile *getActualTile() {return &m_tiles[m_iterator];};
         // SETTERStile
-        void setPosition(unsigned int x, unsigned int y)
+        void setPosition(float x, float y)
         {
             for (auto &tile : m_tiles)
                 tile.setPosition(x, y);
@@ -114,15 +114,15 @@ namespace Arcade
     private:
         std::string m_text;
         Arcade::Color m_color;
-        std::pair<unsigned int, unsigned int> m_pos;
+        std::pair<float, float> m_pos;
     public:
-        explicit Text(std::string text, Arcade::Color color = WHITE, unsigned int x = 0, unsigned int y = 0) : m_text(std::move(text)), m_color(color), m_pos(std::make_pair(x, y)) {};
+        explicit Text(std::string text, Arcade::Color color = WHITE, float x = 0, float y = 0) : m_text(std::move(text)), m_color(color), m_pos(std::make_pair(x, y)) {};
         ~Text() override = default;
         void setText(std::string text) {m_text = std::move(text);};
         void setColor(Arcade::Color color) {m_color = color;};
-        void setPosition(unsigned int x, unsigned int y) {m_pos.first = x; m_pos.second = y;};
+        void setPosition(float x, float y) {m_pos.first = x; m_pos.second = y;};
         [[nodiscard]] std::string getText() const {return m_text;};
         [[nodiscard]] Arcade::Color getColor() const {return m_color;};
-        [[nodiscard]] std::pair<unsigned int, unsigned int> getPosition() const {return m_pos;};
+        [[nodiscard]] std::pair<float, float> getPosition() const {return m_pos;};
     };
 }
